% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atebounds.R
\name{atebounds}
\alias{atebounds}
\title{Bounding the average treatment effect}
\usage{
atebounds(y, t, x, rps, q = 10L, discrete = FALSE)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{rps}{n-dimensional vector of reference propensity scores}

\item{q}{polynomial order (default: q = 10)}

\item{discrete}{TRUE if x inclues only discrete covariates and FALSE if not (default: FALSE)}
}
\value{
An S3 object of type "bter". The object has the following elements.
\item{y1_lb}{the lower bound of E[Y(1)]}
\item{y1_ub}{the upper bound of E[Y(1)]}
\item{y0_lb}{the lower bound of E[Y(0)]}
\item{y0_ub}{the upper bound of E[Y(0)]}
\item{ate_lb}{the lower bound of E[Y(1) - Y(0)]}
\item{att_ub}{the upper bound of E[Y(1) - Y(0)]}
}
\description{
Bounds the average treatment effect under the unconfoundedness assumption without the overlap condition
}
\examples{
# to be added

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
