% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attbounds_discrete_x.R
\name{attbounds_discrete_x}
\alias{attbounds_discrete_x}
\title{Bounding the average treatment effect on the treated with discrete covariates}
\usage{
attbounds_discrete_x(Y, D, X, rps, Q = 2L, studentize = TRUE, small_c = 1e-08)
}
\arguments{
\item{Y}{n-dimensional vector of binary outcomes}

\item{D}{n-dimensional vector of binary treatments}

\item{X}{n by p matrix of covariates}

\item{rps}{n-dimensional vector of reference propensity scores}

\item{Q}{maximal polynomial order that uses the (Q-1) nearest neighbors excluding own observations (default: Q = 2)}

\item{studentize}{TRUE if X is studentized elementwise and FALSE if not (default: TRUE)}

\item{small_c}{a small positive constant to determine the two covariate vectors are identical (default: 1e-8)}
}
\value{
An S3 object of type "ATbounds". The object has the following elements.
\item{att_lb}{the lower bound of ATT, i.e. E[Y(1) - Y(0) | T = 1]}
\item{att_ub}{the upper bound of ATT, i.e. E[Y(1) - Y(0) | T = 1]}
\item{att_rps}{the point estimate of ATT using the reference propensity scores}
}
\description{
Bounds the average treatment effect on the treated under the unconfoundedness assumption without the overlap condition.
This command is for the case when all the covariates are discrete.
}
\examples{
  Y <- RHC[,"survival"]
  D <- RHC[,"RHC"]
  age <- round(RHC[,"age"])
  female <- RHC[,"sex_Female"]
  X <- cbind(age,female)
  rps <- rep(mean(D),length(D))
  results_att <- attbounds_discrete_x(Y, D, X, rps, Q = 3)

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
