% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batt.R
\name{batt}
\alias{batt}
\title{Bounding the average treatment effect on the treated}
\usage{
batt(y, t, x, q = 10L, discrete = FALSE)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{q}{polynomial order (default: q = 10)}

\item{discrete}{TRUE if x inclues only discrete covariates and FALSE if not (default: FALSE)}
}
\value{
An S3 object of type "bter". The object has the following elements.
\item{lb}{the lower bound of ATT}
\item{ub}{the upper bound of ATT}
}
\description{
Bounds the average treatment effect on the treated under the unconfoundedness assumption without the overlap condition
}
\examples{
# to be added

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
