% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atebounds.R
\name{atebounds}
\alias{atebounds}
\title{Bounding the average treatment effect (ATE)}
\usage{
atebounds(Y, D, X, rps, Q = 2L, n_permute = 0L, studentize = TRUE)
}
\arguments{
\item{Y}{n-dimensional vector of binary outcomes}

\item{D}{n-dimensional vector of binary treatments}

\item{X}{n by p matrix of covariates}

\item{rps}{n-dimensional vector of reference propensity scores}

\item{Q}{polynomial order that uses the (Q-1) nearest neighbors excluding own observations (default: Q = 2)}

\item{n_permute}{number of permutations to shuffle the data (default: 0)}

\item{studentize}{TRUE if x is studentized elementwise and FALSE if not (default: TRUE)}
}
\value{
An S3 object of type "ATbounds". The object has the following elements.
\item{y1_lb}{the lower bound of the average of Y(1), i.e. E[Y(1)]}
\item{y1_ub}{the upper bound of the average of Y(1), i.e. E[Y(1)]}
\item{y0_lb}{the lower bound of the average of Y(0), i.e. E[Y(0)]}
\item{y0_ub}{the upper bound of the average of Y(0), i.e. E[Y(0)]}
\item{ate_lb}{the lower bound of ATE, i.e. E[Y(1) - Y(0)]}
\item{ate_ub}{the upper bound of ATE, i.e. E[Y(1) - Y(0)]}
\item{ate_rps}{the point estimate of ATE using the reference propensity scores}
}
\description{
Bounds the average treatment effect (ATE) under the unconfoundedness assumption without the overlap condition
}
\examples{
  Y <- RHC[,"survival"]
  D <- RHC[,"RHC"]
  X <- RHC[,-c(1,2)]
  rps <- rep(mean(D),length(D))  
  results_ate <- atebounds(Y, D, X, rps, Q = 3)

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
