% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attbounds.R
\name{attbounds}
\alias{attbounds}
\title{Bounding the average treatment effect on the treated}
\usage{
attbounds(
  y,
  t,
  x,
  rps,
  q = 2L,
  permute_max = 0,
  discrete = FALSE,
  studentize = TRUE
)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{rps}{n-dimensional vector of reference propensity scores}

\item{q}{polynomial order (default: q = 2, which uses the nearest neighbor excluding own observations)}

\item{permute_max}{maximum number of permutations to shuffle the data (default: 0)}

\item{discrete}{TRUE if x includes only discrete covariates and FALSE if not (default: FALSE)}

\item{studentize}{TRUE if x is studentized elementwise and FALSE if not (default: TRUE)}
}
\value{
An S3 object of type "ATbounds". The object has the following elements.
\item{lb}{the lower bound of ATT, i.e. E[Y(1) - Y(0) | T = 1]}
\item{ub}{the upper bound of ATT, i.e. E[Y(1) - Y(0) | T = 1]}
\item{att_rps}{the point estimate of ATT using the reference propensity score}
}
\description{
Bounds the average treatment effect on the treated under the unconfoundedness assumption without the overlap condition
}
\examples{
# to be added

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
