% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attbounds_discrete_x.R
\name{attbounds_discrete_x}
\alias{attbounds_discrete_x}
\title{Bounding the average treatment effect on the treated (ATT) with discrete covariates}
\usage{
attbounds_discrete_x(Y, D, X, rps, Q = 3L, studentize = TRUE)
}
\arguments{
\item{Y}{n-dimensional vector of binary outcomes}

\item{D}{n-dimensional vector of binary treatments}

\item{X}{n by p matrix of covariates}

\item{rps}{n-dimensional vector of the reference propensity score}

\item{Q}{bandwidth parameter that determines the maximum number of observations for pooling information (default: Q = 3)}

\item{studentize}{TRUE if X is studentized elementwise and FALSE if not (default: TRUE)}
}
\value{
An S3 object of type "ATbounds". The object has the following elements.
\item{att_lb}{estimate of the lower bound on ATT, i.e. E[Y(1) - Y(0) | D = 1]}
\item{att_ub}{estimate of the upper bound on ATT, i.e. E[Y(1) - Y(0) | D = 1]}
\item{att_rps}{the point estimate of ATT using the reference propensity score}
}
\description{
Bounds the average treatment effect on the treated (ATT) under the unconfoundedness assumption without the overlap condition.
This command is for the case when all the covariates are discrete.
}
\examples{
  Y <- RHC[,"survival"]
  D <- RHC[,"RHC"]
  age <- round(RHC[,"age"])
  female <- RHC[,"sex_Female"]
  X <- cbind(age,female)
  rps <- rep(mean(D),length(D))
  results_att <- attbounds_discrete_x(Y, D, X, rps, Q = 3)

}
\references{
Sokbae Lee and Martin Weidner. Bounding Treatment Effects by Pooling Limited Information across Observations.
}
